[ADD] + [FIX]
[ADD] Notification on upload + Click on the notification to see the picture PicShaped
[FIX] Fix back to gallery instead the wall + empty account bug + miscellaneous

+++ app/src/main/java/com/example/android/picshape/view/SinglePicActivity.java
@@ -3,10 +3,10 @@ package com.example.android.picshape.view;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.graphics.drawable.BitmapDrawable;
 import android.os.AsyncTask;
 import android.preference.PreferenceManager;
 import android.provider.MediaStore;
+import android.support.design.widget.FloatingActionButton;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
@@ -20,24 +20,27 @@ import android.widget.Toast;
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.resource.bitmap.GlideBitmapDrawable;
 import com.example.android.picshape.R;
+import com.example.android.picshape.dao.AccountAccess;
 import com.example.android.picshape.dao.AccountSingleton;
-import com.example.android.picshape.dao.PicSingleton;
 import com.example.android.picshape.dao.PictureAccess;
 import com.example.android.picshape.model.PicshapeAccount;
 import com.example.android.picshape.model.PictureShape;
 
-import java.util.ArrayList;
-
 public class SinglePicActivity extends AppCompatActivity {
 
+    private final String MODE_DELETE = "1";
+    private final String MODE_USER = "2";
+
     //View
-    private ImageView mPicture;
+    private ImageView mPicture, mProfilPicture;
     private Button mDeleteButton;
-    private TextView mUsername, mPicTitle;
+    private FloatingActionButton mCommentButton;
+    private TextView mUsernameTv, mPicTitle;
     private PictureShape mPicShape;
     private ProgressBar mDeleteProgressBar;
 
     private boolean converted = true;
+    private PicshapeAccount mPictureAccount;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -51,14 +54,15 @@ public class SinglePicActivity extends AppCompatActivity {
 
         initComp();
 
-        mUsername.setText( intent.getStringExtra("username") );
+        String username = intent.getStringExtra("username");
+
 
         if(mPicShape != null){
             Glide.with(mPicture.getContext())
                     .load(mPicShape.getUrlConverted())
                     .into(mPicture);
 
-            initComp();
+            getPictureAccount(username);
         }
 
     }
@@ -99,14 +103,14 @@ public class SinglePicActivity extends AppCompatActivity {
         });
 
 
-        mUsername = (TextView) findViewById(R.id.name_textView);
+        mUsernameTv = (TextView) findViewById(R.id.name_textView);
 
-        mUsername.setOnClickListener(new View.OnClickListener() {
+        mUsernameTv.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                String name = ((TextView)v).getText().toString();
+
                 Intent galleryIntent = new Intent(SinglePicActivity.this, GalleryActivity.class);
-                galleryIntent.putExtra("account", new PicshapeAccount("-1", name, null, null, null));
+                galleryIntent.putExtra("account", mPictureAccount);
 
                 if (galleryIntent.resolveActivity(getPackageManager()) != null) {
                     startActivity(galleryIntent);
@@ -116,10 +120,11 @@ public class SinglePicActivity extends AppCompatActivity {
 
         mPicTitle = (TextView) findViewById(R.id.pic_name_textView);
 
+        mProfilPicture = (ImageView) findViewById(R.id.profile_pic_imageView);
 
         // If user is the picture owner
 
-        if(mPicShape.getIdUser().equals(AccountSingleton.getInstance().getAccountLoaded().getName())) {
+        if( mPicShape != null && mPicShape.getIdUser().equals(AccountSingleton.getInstance().getAccountLoaded().getName())) {
             mDeleteButton = (Button) findViewById(R.id.delete_button);
 
             mDeleteButton.setVisibility(View.VISIBLE);
@@ -133,25 +138,62 @@ public class SinglePicActivity extends AppCompatActivity {
         }
 
         mDeleteProgressBar = (ProgressBar) findViewById(R.id.single_pic_progressBar);
+
+        mCommentButton = (FloatingActionButton) findViewById(R.id.comment_fab_button);
+
+        mCommentButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                postComment();
+            }
+        });
+    }
+
+    /**
+     * This function display information about the uploader
+     */
+    public void displayUserInfo(){
+        if(mPictureAccount != null) {
+            mUsernameTv.setText(mPictureAccount.getName());
+            Glide.with(mProfilPicture.getContext())
+                    .load(mPictureAccount.getUrlGravatar())
+                    .asBitmap()
+                    .into(mProfilPicture);
+        }
+    }
+
+    /**
+     * This function get the picture's user
+     * @param name
+     */
+    public void getPictureAccount(String name){
+        if(name != null){
+            SinglePicTask task = new SinglePicTask();
+
+            String route = "/api/users/";
+            SharedPreferences preferences =  PreferenceManager.getDefaultSharedPreferences(this);
+            String url = preferences.getString(getString(R.string.pref_url_key),getString(R.string.pref_url_default))+route;
+
+            task.execute(MODE_USER, url, null, name);
+        }
     }
 
     /**
      * This function makes call to delete picture
      */
     public void deletePicture(){
-        DeleteTask task = new DeleteTask();
+        SinglePicTask task = new SinglePicTask();
 
         String url = mPicShape.getUrlConverted();
         String token = AccountSingleton.getInstance().getAccountLoaded().getToken();
 
-        task.execute(url, token);
+        task.execute(MODE_DELETE, url, token, null);
     }
 
     /**
      * This function post a comment to the picture
-     * @param v
      */
-    public void postComment(View v){
+    public void postComment(){
         //TODO it miss comment function on the back
         showText("Comment feature soon available");
     }
@@ -204,7 +246,6 @@ public class SinglePicActivity extends AppCompatActivity {
         builder.show();
     }
 
-
     /**
      * This function saves picture in internal memory
      * @return
@@ -225,8 +266,10 @@ public class SinglePicActivity extends AppCompatActivity {
     /**
      * This class manages the http call to webService
      */
-    class DeleteTask extends AsyncTask<String, Void, Integer> {
+    class SinglePicTask extends AsyncTask<String, Void, Integer> {
 
+        private PicshapeAccount mUserAccount;
+        private String mMode;
 
         /**
          * This function is executed before execution of the task, used to show progress bar
@@ -246,13 +289,22 @@ public class SinglePicActivity extends AppCompatActivity {
         protected Integer doInBackground(String[] params){
 
             if(params.length < 2) return -1;
+            mMode = params[0];
+            String url = params[1];
+            String token = params[2];
+            String name = params[3];
 
-            String url = params[0];
-            String token = params[1];
-
-            if( PictureAccess.deletePicture(url, token) ) return 0;
-            else return 1;
-
+            if(MODE_DELETE.equals(mMode)){
+                if( PictureAccess.deletePicture(url, token) ) return 0;
+            }
+            else if(MODE_USER.equals(mMode)){
+                mUserAccount = AccountAccess.getUserByName(url, name);
+                if( mUserAccount != null ) return 0;
+            }
+            else{
+                return 1;
+            }
+            return 1;
         }
 
 
@@ -265,7 +317,14 @@ public class SinglePicActivity extends AppCompatActivity {
             stopLoading();
             if(result == 0){
                 // Success
-                launchGallery();
+                if(MODE_DELETE.equals(mMode)){
+                    launchGallery();
+                }
+                else if(MODE_USER.equals(mMode)){
+                    mPictureAccount = mUserAccount;
+                    displayUserInfo();
+                }
+
             }
             else{
                 showText("Error, try again");
